+++ controllers/mybooks/index.js
@@ -1,0 +1,6 @@
+const router = require('express').Router();
+const myBooksRoutes = require('./myBooksRoutes');
+
+router.use('/', myBooksRoutes);
+
+module.exports = router;
